Refactor Version command
