+++ src/commands/mod.rs
@@ -8,5 +8,4 @@ pub mod chunk_get_rsp;
 pub mod digests_rsp;
 pub mod error_rsp;
 pub mod measurements_rsp;
-pub mod version_rq;
-pub mod version_rsp;
+pub mod version;
